chore(schemas): move database definition files from schemas/ to
