+++ packages/internal/schemas/drizzle.config.ts
@@ -2,6 +2,6 @@ import { defineConfig } from "drizzle-kit";
 
 export default defineConfig({
 	out: "./src/drizzle",
-	schema: "./src/schemas/*.ts",
+	schema: "./src/schemas/db/*.ts",
 	dialect: "postgresql",
 });
