+++ packages/internal/schemas/src/schemas/db/architectures.ts
